+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Automatically generate examples for apply command
 - Update NPM keywords
 
 ## [0.5.0] - 2020-11-02